refactor(app.models): make base_model and task_model instead of models
