+++ alembic/env.py
@@ -8,8 +8,8 @@ from sqlalchemy.ext.asyncio import async_engine_from_config
 from alembic import context
 from app.config import config as env
 
-from app.models.models import *
-from app.models.models import Base
+from app.models import User, Task
+from app.models.base_model import Base
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
 config = context.config
