+++ app/models/__init__.py
@@ -1,2 +1,7 @@
-from .models import *
-from .models import User, Task, Base
+from app.models.user_model import User
+from app.models.task_model import Task
+
+__all__ = [
+    'User',
+    'Task'
+]
