+++ app/models/base_model.py
@@ -1,0 +1,5 @@
+from sqlalchemy.orm import DeclarativeBase
+
+class Base(DeclarativeBase):
+    pass
+
